encoding: Move DLSettings from maccommands to types module
