+++ lorawan-encoding/tests/lorawan.rs
@@ -4,7 +4,7 @@ use lorawan::keys::*;
 use lorawan::maccommandcreator::*;
 use lorawan::maccommands::*;
 use lorawan::parser::*;
-use lorawan::types::Frequency;
+use lorawan::types::{DLSettings, Frequency};
 
 fn phy_join_request_payload() -> Vec<u8> {
     let mut res = Vec::new();
